* subversion/tests/libsvn_client/client-test.c   (check_patch_result):     Following up on r927785, cast strlen() result to int to fix some     64 bit warnings.
Found by: philip's compiler


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@927793 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/tests/libsvn_client/client-test.c
@@ -232,8 +232,8 @@ check_patch_result(const char *path, const char **expected_lines,
           return svn_error_createf(SVN_ERR_TEST_FAILED, NULL,
                                    "%s line %d didn't match the expected line "
                                    "(strlen=%d vs strlen=%d)", path, i,
-                                   strlen(expected_lines[i-1]),
-                                   strlen(line->data));
+                                   (int)strlen(expected_lines[i-1]),
+                                   (int)strlen(line->data));
 
       if (eof)
         break;
